requirements_master.txt
flask
pandas
Flask-Cors
django~=5.2
tabulate~=0.9
python-dotenv
psycopg2-binary
whitenoise~=6.9
django-ipware~=7.0
django-timescaledb~=0.2
djangorestframework~=3.16
django-cors-headers~=4.7.0